+++ src/fx/run_once.rs
@@ -1,11 +1,8 @@
-use alloc::{boxed::Box, vec};
+use alloc::boxed::Box;
 
 use ratatui_core::{buffer::Buffer, layout::Rect};
 
-use crate::{
-    effect::Effect, shader::Shader, widget::EffectSpan, CellFilter, ColorSpace, Duration,
-    EffectTimer,
-};
+use crate::{effect::Effect, shader::Shader, CellFilter, ColorSpace, Duration, EffectTimer};
 
 /// A shader that wraps another effect and ensures it runs exactly once before reporting
 /// completion.
@@ -77,10 +74,6 @@ impl Shader for RunOnce {
     fn reset(&mut self) {
         self.has_run = false;
         self.effect.reset();
-    }
-
-    fn as_effect_span(&self, offset: Duration) -> EffectSpan {
-        EffectSpan::new(self, offset, vec![self.effect.as_effect_span(offset)])
     }
 
     fn color_space(&self) -> ColorSpace {